Fix and Refactor Newsletter model
Add summary field to Paper that will be displayed in the newsletter email sent to customers instead of the abstract.
Link the newsletter to a Paper Topic instead of a subject and a ManyToManyField to Subscriber so a subscriber can subscribe to multiple Newsletter and a Newsletter can be accessed by multiple subscribers.
Add PersonalNewsletter models to enables user create a custom newsletter of their preferred topic/sentence/keywords